Fix: Type assertion for total_points from int64 to float64 in usage checker (#103)

+++ apps/backend/internal/services/usagechecker.go
@@ -138,7 +138,7 @@ func (uc *UsageChecker) getCurrentDailyUsage(ctx context.Context, userID string)
 	var totalPoints int
 	for _, doc := range docs {
 		data := doc.Data()
-		if points, ok := data["total_points"].(int64); ok {
+		if points, ok := data["total_points"].(float64); ok {
 			totalPoints += int(points)
 		}
 	}
